Multi-Objective Crystal Structure Algorithm (MOCryStAl)
In many optimization problems, the main goal is to improve a single performance index in which a minimum or maximum value of this index fully reflects the quality of the response obtained from a system. However, in some cases, it is impossible to rely solely on a single index, so a multi-objective optimization problem with multiple performance indicators is considered where the values of all of them should be optimized simultaneously. The mentioned process requires a multi-objective optimization algorithm that can deal with the complexity of problems with simultaneous indexes. This paper presents the multi-objective version of a recently proposed metaheuristic algorithm called Crystal Structure Algorithm (CryStAl) which was inspired by the principles underlying the formation crystal structures. For the performance evaluation of this algorithm which is called MOCryStAl, the benchmark problems of the Completions on Evolutionary Computation (CEC) on multi- objective optimization, called CEC-09, are utilized. Some real-world engineering design problems are used to evaluate the efficiency of the proposed approach. The results demonstrate that the proposed methods can provide excellent results in dealing with the considered multi-objective problems.